Modularize the vehicle component
